refactored code. created a new component for GenerateStoryPage, and imported that into App.js
