+++ package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.7.4",
     "react-bootstrap-icons": "^1.10.3",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "typewriter-effect": "^2.19.0",
     "web-vitals": "^2.1.4"
